+++ ConcreteApplicationApp/Protocols/CodeView.swift
@@ -1,0 +1,27 @@
+//
+//  CodeView.swift
+//  ConcreteApplication
+//
+//  Created by Bruno Cruz on 20/12/18.
+//  Copyright © 2018 Bruno Cruz. All rights reserved.
+//
+
+import Foundation
+
+protocol CodeView{
+    func buildViewHierarchy()
+    func setupConstraints()
+    func setupAdditionalConfiguration()
+    func setupView()
+}
+
+extension CodeView{
+    func setupView(){
+        buildViewHierarchy()
+        setupConstraints()
+        setupAdditionalConfiguration()
+    }
+    
+    func setupAdditionalConfiguration() {}
+    
+}
